Fixed multiple glitches.

+++ resources/spikey-ball/ball-script.lua
@@ -1,6 +1,11 @@
 events.trigger{
     collision = function(self, event)
-        local collider = event:getCollider()
-        common.kill(collider, true)
+        -- Only after being dropped
+        if self:isCollisionChecker() then
+            local collider = event:getCollider()
+            common.kill(collider, true)
+            effects.stars(self:getPosition(), 5)
+            self:setGarbage()
+        end
     end
 }
